Delay second buffer initialisation.
We don't actually need to initialise our second Cairo context / surface
until after the first one has been painted to the screen. This commit
therefore delays this initialisation (or at least the expensive memcpy),
granting a significant reduction in startup time.

The downside is that main() and entry_init() are now tied together
somewhat, but hopefully the comments help.