Abandoned initial work on damage tracking.
This is working for the Harfbuzz backend at least, but I'm going to
shelve it (at least for now). The end result isn't really any faster
than just redrawing the whole screen every frame (at least on tofi's
side), and it adds a significant amount of complexity. Still, it's valid
work, and may come in useful in the future.