Add Levenshtein distance implementation.
This works, but isn't integrated properly with the rest of the codebase.

Having played around with it for quite a while, however, I've decided
that Levenshtein distance is just a pretty rubbish metric for something
like tofi. You have to have typed something very close to the *whole*
target string to get the result you want, and it doesn't play very nice
with the history score.

I'm therefore abandoning this for now.